Finished the proyect We have tested for bugs, correct working of the proyect, functionalities, etc, and all seems pretty good, therefore we can conclude the development. Still, we are going to test a little further to search for potential problems.
