--- backEnd/DecompressPageBackEnd.java
@@ -1,4 +1,0 @@
-package backEnd;
-
-public class DecompressPageBackEnd {
-}
