--- backEnd/FileExtensionException.java
@@ -1,10 +1,0 @@
-package backEnd;
-
-import java.rmi.server.ExportException;
-
-public class FileExtensionException extends Exception {
-    public FileExtensionException()
-    {
-        super("The app does not support this file extension");
-    }
-}
