--- backEnd/MainPageBackEnd.java
@@ -1,5 +1,0 @@
-package backEnd;
-
-public class MainPageBackEnd {
-
-}
